+++ NEWS.md
@@ -2,9 +2,9 @@
 
 ### v0.10.3
 
-* `trakt.user.stats`:
-        * Tidy up rating distribution 
-        * **Remove** `to.data.frame` option because the output is too messy
+* `trakt.user.stats`:  
+    * Tidy up rating distribution 
+    * **Remove** `to.data.frame` option because the output is too messy
 * Add another `@family` tag to docs for aggregation functions (`.popular`, `.trending`, `.related`)
 
 ### v0.10.2