+++ NEWS.md
@@ -7,6 +7,9 @@
     * **Remove** `to.data.frame` option because the output is too messy
 * Add another `@family` tag to docs for aggregation functions (`.popular`, `.trending`, `.related`)
 * Individual functions don't have to warn about missing headers, that's `trakt.api.call`‘s job.
+* `trakt.getEpisodeData`: 
+    * Explicitly drop episodes with NA firstaired fields
+    * The `episode_abs` field is usually `NA`, so let's dump `epnum` on it
 
 ### v0.10.2